+++ CHANGELOG.md
@@ -4,6 +4,9 @@ It covers all changes made from v2.0.4 onwards
 
 ## [Unreleased]
 
+### Fixed
+- Made blank *_cosmic.csv and *_intersect.txt required as AutoQC markers
+
 ## [v2.1.0] - 2024-10-03
 
 ### Added